renamed files, better file structuring of new components
