--- docs/achievement/card.mdx
@@ -1,30 +1,0 @@
----
-name: AchievementCard
-route: /achievement/card
-menu: Achievement
-edit: false
----
-
-import { Playground, Props } from "docz"
-import { AchievementCard } from "../../src/google-classroom"
-import { AchievementSuperGiver, AchievementSharer, AchievementFoodRun } from "./data"
-
-# Achievement Box Component
-
-Use this component in Embedded views
-
-<Playground>
-	<div style={{ display: "flex", justifyContent: "space-evenly" }}>
-		<div style={{ width: "20%", display: "flex", justifyContent: "center" }}>
-			<AchievementCard data={AchievementFoodRun} openDetails={(data) => console.log(data)} />
-		</div>
-		<div style={{ width: "20%", display: "flex", justifyContent: "center" }}>
-			<AchievementCard data={AchievementSharer} openDetails={(data) => console.log(data)} />
-		</div>
-		<div style={{ width: "20%", display: "flex", justifyContent: "center" }}>
-			<AchievementCard data={AchievementSuperGiver} openDetails={(data) => console.log(data)} />
-		</div>
-	</div>
-</Playground>
-
-<Props of={AchievementCard} />
