+++ docs/achievement/details.mdx
@@ -1,17 +1,17 @@
 ---
-name: AchievementDetails
+name: Details
 route: /achievement/details
 menu: Achievement
 edit: false
 ---
 
 import { Playground, Props } from "docz"
-import { AchievementDetails } from "../../src/google-classroom"
+import { AchievementDetails } from "../../src"
 import { AchievementSharer } from "./data"
 
-# Achievement Box Component
+# Achievement Details Component
 
-Use this component in Embedded views
+Use this component in to display details about a selected achievement
 
 <Playground>
     <div style={{display: "flex", alignItems: "center", justifyContent: "center", flexDirection: "column", height: "700px"}}>
