+++ docs/character/progress-circular.mdx
@@ -6,11 +6,11 @@ edit: false
 ---
 
 import { Playground, Props } from "docz"
-import { ProgressCircle } from "../../src/google-classroom"
+import { ProgressCircle } from "../../src"
 
 # Progress Circle Component
 
-Use this component to ....
+Use this component to display character progressions through a circular progress bar
 
 <Playground>
 	<div style={{ display: "flex", alignItems: "center", justifyContent: "space-around" }}>
