+++ docs/character/tiered-progress-bar.mdx
@@ -6,12 +6,12 @@ edit: false
 ---
 
 import { Playground, Props } from "docz"
-import { TieredProgressBar } from "../../src/google-classroom"
+import { TieredProgressBar } from "../../src"
 import { Tiers1, Tiers2 } from "./data"
 
 # Tiered Progress Bar Component
 
-Use this component to ....
+Use this component to display detailed character progression
 
 <Playground>
 	<div style={{ display: "flex", flexDirection: "column", height: "200px", width: "400px", justifyContent: "space-around" }}>
