+++ docs/shop/details.mdx
@@ -1,12 +1,12 @@
 ---
-name: ShopItemDetails
+name: Item Details
 route: /shop/details
 menu: Shop
 edit: false
 ---
 
 import { Playground, Props } from "docz"
-import { ShopItemDetails } from "../../src/google-classroom"
+import { ShopItemDetails } from "../../src"
 import { ShopItemBoost, ShopItemShirt, ModuleShop } from "./data"
 
 # Shop Item Details Component
