+++ src/character/index.ts
