--- src/google-classroom/achievement/index.ts
@@ -1,2 +1,0 @@
-export * from "./achievement-card"
-export * from "./achievement-details"
