--- src/google-classroom/components/rewards/index.ts
@@ -1,1 +1,0 @@
-export * from "./rewards"
