--- src/google-classroom/index.ts
@@ -1,3 +1,0 @@
-export * from "./achievement"
-export * from "./character"
-export * from "./shop"
