--- src/google-classroom/shop/index.ts
@@ -1,2 +1,0 @@
-export * from "./shop-item-card"
-export * from "./shop-item-details"
