+++ src/index.ts
@@ -1,7 +1,9 @@
 export * from "./achievement"
+export * from "./character"
 export * from "./components"
 export * from "./leaderboard"
 export * from "./level"
 export * from "./mission"
+export * from "./shop"
 export * from "./theme/options"
 export * from "./theme/theme"
