+++ src/shop/item-details.tsx
@@ -1,8 +1,8 @@
 import React from "react"
 import styled from "styled-components"
-import { rewardIcons } from "../../icons/rewards"
-import { ModuleShopItem, ShopData } from "../../types"
-import { Color } from "../../utils/color"
+import { rewardIcons } from "../icons/rewards"
+import { ModuleShopItem, ShopData } from "../types"
+import { Color } from "../utils/color"
 import { userCanAfford } from "./utils/can-afford"
 import { getShopIcon } from "./utils/icon"
 import { parseShopCurrency } from "./utils/parse-currency"
