+++ src/shop/utils/can-afford.ts
@@ -1,4 +1,4 @@
-import { ModuleShopItem, ShopData } from "../../../types"
+import { ModuleShopItem, ShopData } from "../../types"
 
 export function userCanAfford(item: ModuleShopItem, shop: ShopData) {
 	return item.currency === "vip" ? !!shop.vipCurrency && item.value <= shop.vipCurrency : item.value <= shop.coins
