+++ src/shop/utils/parse-currency.ts
@@ -1,4 +1,4 @@
-import { Currency, ShopCurrency } from "../../../types"
+import { Currency, ShopCurrency } from "../../types"
 
 export function parseShopCurrency(currency: ShopCurrency): Currency {
 	return currency === "vip" ? "vipcurrency" : "coin"
